requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.11.3
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
alsa-lib=1.2.14=h5eee18b_0
aom=3.6.0=h6a678d5_0
asteval=1.0.6=pyhd8ed1ab_0
blas=2.136=openblas
blas-devel=3.9.0=36_h1ea3ea9_openblas
blosc=1.21.6=hf7d4471_0
bottleneck=1.4.2=py39ha9d4c09_0
brotli-python=1.0.9=py39h6a678d5_9
bzip2=1.0.8=h5eee18b_6
c-ares=1.34.5=hef5626c_0
c-blosc2=2.17.1=h4257ac4_0
ca-certificates=2025.9.9=h06a4308_0
cairo=1.18.4=h44eff21_0
contourpy=1.2.1=py39hdb19cb5_1
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h1110e0f_3
dav1d=1.2.1=h5eee18b_0
dbus=1.16.2=h5bd4931_0
dill=0.3.8=py39h06a4308_0
expat=2.7.1=h6a678d5_0
fontconfig=2.14.1=h55d465d_3
fonttools=4.55.3=py39h5eee18b_0
freetype=2.13.3=h4a9f257_0
fretbursts=0.8.3=py39hf15bc26_1
fribidi=1.0.10=h7b6447c_0
future=1.0.0=py39h06a4308_0
graphite2=1.3.14=h295c915_1
harfbuzz=10.2.0=hdfddeaa_1
hdf5=1.14.5=h2b7332f_2
icu=73.1=h6a678d5_0
importlib-metadata=8.5.0=py39h06a4308_0
importlib_resources=6.4.0=py39h06a4308_0
jpeg=9f=h5ce9db8_0
kiwisolver=1.4.4=py39h6a678d5_0
lcms2=2.16=h92b89f2_1
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h6a678d5_0
libabseil=20250127.0=cxx17_h6a678d5_0
libavif=1.1.1=h5eee18b_0
libblas=3.9.0=36_h4a7cf45_openblas
libcblas=3.9.0=36_h0358290_openblas
libclang13=20.1.8=default_hee3e7a4_0
libcups=2.4.2=h252cb56_2
libcurl=8.15.0=hc1efc7f_0
libdeflate=1.22=h5eee18b_0
libdrm=2.4.124=h5eee18b_0
libegl=1.7.0=h5eee18b_2
libev=4.33=h7f8727e_1
libevent=2.1.12=hdbd6064_1
libffi=3.4.4=h6a678d5_1
libgcc=15.1.0=h767d61c_5
libgcc-ng=15.1.0=h69a702a_5
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgl=1.7.0=h5eee18b_2
libglib=2.84.2=h37c7471_0
libglvnd=1.7.0=h5eee18b_2
libglx=1.7.0=h5eee18b_2
libgomp=15.1.0=h767d61c_5
libiconv=1.16=h5eee18b_3
libkrb5=1.21.3=h520c7b4_4
liblapack=3.9.0=36_h47877c9_openblas
liblapacke=3.9.0=36_h6ae95b6_openblas
libllvm15=15.0.7=he89c38a_4
libllvm20=20.1.8=h465586c_0
libmpdec=4.0.0=h5eee18b_0
libnghttp2=1.57.0=h2d74bed_0
libopenblas=0.3.30=h46f56fc_0
libopengl=1.7.0=h5eee18b_2
libopus=1.3.1=h5eee18b_1
libpciaccess=0.18=h5eee18b_0
libpng=1.6.39=h5eee18b_0
libpq=17.4=h02b6914_2
libre2-11=2024.07.02=h6a678d5_0
libsodium=1.0.20=heac8642_0
libssh2=1.11.1=h251f7ec_0
libstdcxx=15.1.0=h8f9b012_5
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.7.0=hde9077f_0
libuuid=1.41.5=h5eee18b_0
libvpx=1.13.1=h6a678d5_0
libwebp-base=1.3.2=h5eee18b_1
libxcb=1.17.0=h9b100fa_0
libxkbcommon=1.9.1=h69220b7_0
libxkbfile=1.1.0=hf54ef8a_2
libxml2=2.13.8=hfdd30dd_0
libxslt=1.1.41=h097e994_0
libzlib=1.3.1=hb25bd0a_0
lmdb=0.9.31=hb25bd0a_0
lmfit=1.3.4=pyhd8ed1ab_0
lz4-c=1.9.4=h6a678d5_1
lzo=2.10=h7b6447c_2
matplotlib-base=3.9.2=py39hbfdbfaf_1
mesalib=25.1.5=h31e3550_0
mysql-common=9.3.0=h0021b1a_3
mysql-libs=9.3.0=h797a5cd_3
ncurses=6.5=h7934f7d_0
nodegraphqt=0.6.42=pypi_0
nomkl=3.0=0
nspr=4.35=h6a678d5_0
nss=3.89.1=h6a678d5_0
numexpr=2.10.1=py39hd28fd6d_0
numpy=1.26.4=py39hc213724_1
numpy-base=1.26.4=py39hf387b04_1
openblas=0.3.30=h5426c36_0
openblas-devel=0.3.30=h557178c_0
openjpeg=2.5.2=h0d4d230_1
openldap=2.6.10=he9288cc_1
openssl=3.5.3=h26f9b46_1
packaging=25.0=py39h06a4308_0
pandas=2.3.2=py39h277b779_0
pathlib=1.0.1=pyhd3eb1b0_1
pcre2=10.42=hebb0a14_1
phconvert=0.9.1=pyhd8ed1ab_0
pillow=11.3.0=py39hb1c3d2d_0
pip=25.2=pyhc872135_0
pixman=0.46.4=h7934f7d_0
ply=3.11=py39h06a4308_0
pthread-stubs=0.3=h0ce48e5_1
py-cpuinfo=9.0.0=py39h06a4308_0
pybind11-abi=4=hd3eb1b0_1
pyparsing=3.2.0=py39h06a4308_0
pyqt=6.9.1=py39hc79087c_0
pyqt5=5.15.11=pypi_0
pyqt5-qt5=5.15.17=pypi_0
pyqt5-sip=12.17.0=pypi_0
pyqt6-sip=13.10.2=py39hee96239_0
pyside6=6.9.2=py39h2ddcaae_0
pytables=3.9.2=py39h1bd3de6_1
python=3.9.23=he99959a_0
python-dateutil=2.9.0post0=py39h06a4308_2
python-tzdata=2025.2=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytz=2025.2=py39h06a4308_0
qtbase=6.9.2=h46e0f57_0
qtdeclarative=6.9.2=h55666de_0
qtpy=2.4.3=py39h06a4308_0
qtshadertools=6.9.2=hbdd6827_0
qtsvg=6.9.2=h75a08ff_0
qttools=6.9.2=hc856bc8_0
qtwayland=6.9.2=h9f1b7ec_0
qtwebchannel=6.9.2=hbdd6827_0
qtwebengine=6.9.1=hfb9d843_0
qtwebsockets=6.9.2=hbdd6827_0
re2=2024.07.02=hdb19cb5_0
readline=8.3=hc2a1206_0
scipy=1.13.1=py39heeff2f4_1
seaborn=0.13.2=py39h06a4308_3
setuptools=72.1.0=py39h06a4308_0
setuptools-scm=8.3.1=py39h06a4308_1
sip=6.12.0=py39hf88036b_0
six=1.17.0=py39h06a4308_0
snappy=1.2.1=h6a678d5_0
spirv-tools=2025.1=hdb19cb5_0
sqlite=3.50.2=hb25bd0a_1
tk=8.6.15=h54e0aa7_0
tomli=2.2.1=py39h06a4308_0
types-pyside2=5.15.2.1.7=pypi_0
typing-extensions=4.15.0=py39h06a4308_0
typing_extensions=4.15.0=py39h06a4308_0
tzdata=2025b=h04d1e81_0
uncertainties=3.2.3=pyhd8ed1ab_0
unicodedata2=15.1.0=py39h5eee18b_1
wayland=1.24.0=hdac8c69_0
wheel=0.45.1=py39h06a4308_0
xcb-util=0.4.1=h5eee18b_2
xcb-util-cursor=0.1.5=h5eee18b_0
xcb-util-image=0.4.0=h5eee18b_2
xcb-util-keysyms=0.4.1=h5eee18b_0
xcb-util-renderutil=0.3.10=h5eee18b_0
xcb-util-wm=0.4.2=h5eee18b_0
xkeyboard-config=2.44=h5eee18b_0
xorg-libice=1.1.2=h9b100fa_0
xorg-libsm=1.2.6=h9b100fa_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxcomposite=0.4.6=h9b100fa_2
xorg-libxdamage=1.1.6=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-libxext=1.3.6=h9b100fa_0
xorg-libxfixes=6.0.1=h9b100fa_0
xorg-libxi=1.8.2=h9b100fa_0
xorg-libxrandr=1.5.4=h9b100fa_0
xorg-libxrender=0.9.12=h9b100fa_0
xorg-libxshmfence=1.3.3=h9b100fa_0
xorg-libxtst=1.2.5=h9b100fa_3
xorg-libxxf86vm=1.1.6=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.6.4=h5eee18b_1
zipp=3.21.0=py39h06a4308_0
zlib=1.3.1=hb25bd0a_0
zlib-ng=2.0.7=h5eee18b_0
zstd=1.5.7=h11fc155_0
